fix: ADR-0002 compliance for policy enforcement
All show and delete handlers now correctly pass input.target=null and
input.existing={resource} instead of passing the resource as target.

Handler changes (v3 + v4):
- user: show, delete, groups
- role: show, delete (v4 delegates to v3)
- group: show, delete (v4 delegates to v3)
- token restriction: show, delete (v4-only)

Rego changes:
- identity.rego: global_role/own_role now read from both target and
  existing
- 8 policy docs updated to reflect input.existing for show/delete
- token restriction delete: reads input.existing.restriction.user_id

v4 user: split monolithic mod.rs into per-operation submodules.
