+++ crates/keystone/src/api/v4/user/mod.rs
@@ -11,578 +11,55 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
+//! User API.
+//!
+//! OPA input structure for all user operations:
+//! ```text
+//! {
+//!   "credentials": { ... },
+//!   "target": { "user": <object-or-null> },
+//!   "existing": { "user": <object-or-null> }
+//! }
+//! ```
+//!
+//! | Operation | `input.target.user`  | `input.existing.user` |
+//! |-----------|-----------------------|------------------------|
+//! | Create    | UserCreate payload    | null                   |
+//! | Update    | UserUpdate patch      | current user           |
+//! | Show      | null                  | current user           |
+//! | Delete    | null                  | current user           |
+//! | List      | UserListParameters    | null                   |
 
-use axum::{
-    Json,
-    extract::{Path, Query, State},
-    http::StatusCode,
-    response::IntoResponse,
-};
+use utoipa::OpenApi;
 use utoipa_axum::{router::OpenApiRouter, routes};
 
-use crate::api::auth::Auth;
-use crate::api::error::KeystoneApiError;
-use crate::api::v3::group::types::{Group, GroupList};
 use crate::keystone::ServiceState;
-use openstack_keystone_core::auth::ExecutionContext;
-use types::{User, UserCreateRequest, UserList, UserListParameters, UserResponse};
 
 pub mod types;
 
+mod create;
+mod delete;
+mod groups;
+mod list;
+mod show;
+mod update;
+
+/// OpenApi specification for the user api.
+#[derive(OpenApi)]
+#[openapi(
+    tags(
+        (name="users", description=r#"User API.
+
+User management endpoints for creating, retrieving, updating, and deleting users.
+        "#),
+    )
+)]
+pub struct ApiDoc;
+
 pub(super) fn openapi_router() -> OpenApiRouter<ServiceState> {
     OpenApiRouter::new()
-        .routes(routes!(list, create))
-        .routes(routes!(show, remove))
-        .routes(routes!(groups))
-}
-
-/// List users
-#[utoipa::path(
-    get,
-    path = "/",
-    params(UserListParameters),
-    description = "List users",
-    responses(
-        (status = OK, description = "List of users", body = UserList),
-        (status = 500, description = "Internal error", example = json!(KeystoneApiError::InternalError(String::from("id = 1"))))
-    ),
-    tag="users"
-)]
-#[tracing::instrument(name = "api::user_list", level = "debug", skip(state))]
-async fn list(
-    Auth(user_auth): Auth,
-    Query(query): Query<UserListParameters>,
-    State(state): State<ServiceState>,
-) -> Result<impl IntoResponse, KeystoneApiError> {
-    let users: Vec<User> = state
-        .provider
-        .get_identity_provider()
-        .list_users(
-            &ExecutionContext::from_auth(&state, &user_auth),
-            &query.into(),
-        )
-        .await?
-        .into_iter()
-        .map(Into::into)
-        .collect();
-    Ok((StatusCode::OK, Json(UserList { users })).into_response())
-}
-
-/// Get single user
-#[utoipa::path(
-    get,
-    path = "/{user_id}",
-    params(),
-    responses(
-        (status = OK, description = "Single user", body = UserResponse),
-        (status = 404, description = "User not found", example = json!(KeystoneApiError::NotFound(String::from("id = 1"))))
-    ),
-    tag="users"
-)]
-#[tracing::instrument(name = "api::user_get", level = "debug", skip(state))]
-async fn show(
-    Auth(user_auth): Auth,
-    Path(user_id): Path<String>,
-    State(state): State<ServiceState>,
-) -> Result<impl IntoResponse, KeystoneApiError> {
-    Ok((
-        StatusCode::OK,
-        Json(UserResponse {
-            user: User::from(
-                state
-                    .provider
-                    .get_identity_provider()
-                    .get_user(&ExecutionContext::from_auth(&state, &user_auth), &user_id)
-                    .await
-                    .map(|x| {
-                        x.ok_or_else(|| KeystoneApiError::NotFound {
-                            resource: "user".into(),
-                            identifier: user_id,
-                        })
-                    })??,
-            ),
-        }),
-    )
-        .into_response())
-}
-
-/// Create user
-#[utoipa::path(
-    post,
-    path = "/",
-    description = "Create new user",
-    responses(
-        (status = CREATED, description = "New user", body = UserResponse),
-    ),
-    tag="users"
-)]
-#[tracing::instrument(name = "api::create_user", level = "debug", skip(state))]
-async fn create(
-    Auth(user_auth): Auth,
-    Query(query): Query<UserListParameters>,
-    State(state): State<ServiceState>,
-    Json(req): Json<UserCreateRequest>,
-) -> Result<impl IntoResponse, KeystoneApiError> {
-    let user = state
-        .provider
-        .get_identity_provider()
-        .create_user(&ExecutionContext::from_auth(&state, &user_auth), req.into())
-        .await?;
-    Ok((
-        StatusCode::CREATED,
-        Json(UserResponse {
-            user: User::from(user),
-        }),
-    )
-        .into_response())
-}
-
-/// Delete user
-#[utoipa::path(
-    delete,
-    path = "/{user_id}",
-    description = "Delete user by ID",
-    params(),
-    responses(
-        (status = 204, description = "Deleted"),
-        (status = 404, description = "User not found", example = json!(KeystoneApiError::NotFound(String::from("id = 1"))))
-    ),
-    tag="users"
-)]
-#[tracing::instrument(name = "api::user_delete", level = "debug", skip(state))]
-async fn remove(
-    Auth(user_auth): Auth,
-    Path(user_id): Path<String>,
-    State(state): State<ServiceState>,
-) -> Result<impl IntoResponse, KeystoneApiError> {
-    state
-        .provider
-        .get_identity_provider()
-        .delete_user(&ExecutionContext::from_auth(&state, &user_auth), &user_id)
-        .await?;
-    Ok((StatusCode::NO_CONTENT).into_response())
-}
-
-/// List groups a user is member of
-#[utoipa::path(
-    get,
-    path = "/{user_id}/groups",
-    description = "List groups a user is member of",
-    responses(
-        (status = OK, description = "List of user groups", body = GroupList),
-        (status = 500, description = "Internal error", example = json!(KeystoneApiError::InternalError(String::from("id = 1"))))
-    ),
-    tag="users"
-)]
-#[tracing::instrument(name = "api::user_list", level = "debug", skip(state))]
-async fn groups(
-    Auth(user_auth): Auth,
-    Path(user_id): Path<String>,
-    State(state): State<ServiceState>,
-) -> Result<impl IntoResponse, KeystoneApiError> {
-    let groups: Vec<Group> = state
-        .provider
-        .get_identity_provider()
-        .list_groups_of_user(&ExecutionContext::from_auth(&state, &user_auth), &user_id)
-        .await?
-        .into_iter()
-        .map(Into::into)
-        .collect();
-    Ok((StatusCode::OK, Json(GroupList { groups })).into_response())
-}
-
-#[cfg(test)]
-mod tests {
-    use axum::{
-        body::Body,
-        http::{self, Request, StatusCode},
-    };
-    use http_body_util::BodyExt; // for `collect`
-    use tower::ServiceExt; // for `call`, `oneshot`, and `ready`
-    use tower_http::trace::TraceLayer;
-
-    use openstack_keystone_core::auth::ExecutionContext;
-    use openstack_keystone_core_types::identity::*;
-
-    use super::openapi_router;
-    use crate::api::tests::{get_mocked_state, test_fixture_scoped};
-    use crate::api::v3::group::types::{GroupBuilder as ApiGroupBuilder, GroupList};
-    use crate::api::v3::user::types::{
-        UserBuilder as ApiUser, UserCreateBuilder as ApiUserCreate, UserCreateRequest, UserList,
-        UserResponse as ApiUserResponse,
-    };
-    use crate::identity::{
-        MockIdentityProvider,
-        error::IdentityProviderError,
-        //    types::{Group, UserCreate, UserListParameters, UserResponseBuilder},
-    };
-    use crate::provider::Provider;
-
-    #[tokio::test]
-    async fn test_list() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_list_users()
-            .withf(|_, _: &UserListParameters| true)
-            .returning(|_, _| {
-                Ok(vec![
-                    UserResponseBuilder::default()
-                        .id("1")
-                        .domain_id("did")
-                        .enabled(true)
-                        .name("2")
-                        .build()
-                        .unwrap(),
-                ])
-            });
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .uri("/")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let res: UserList = serde_json::from_slice(&body).unwrap();
-        assert_eq!(
-            vec![
-                ApiUser::default()
-                    .id("1")
-                    .domain_id("did")
-                    .name("2")
-                    .enabled(true)
-                    .build()
-                    .unwrap(),
-            ],
-            res.users
-        );
-    }
-
-    #[tokio::test]
-    async fn test_list_qp() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_list_users()
-            .withf(|_, qp: &UserListParameters| {
-                UserListParameters {
-                    domain_id: Some("domain".into()),
-                    name: Some("name".into()),
-                    ..Default::default()
-                } == *qp
-            })
-            .returning(|_, _| Ok(Vec::new()));
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .uri("/?domain_id=domain&name=name")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let _res: UserList = serde_json::from_slice(&body).unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_list_unauth() {
-        let state = get_mocked_state(Provider::mocked_builder(), false, None).await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state);
-
-        let response = api
-            .as_service()
-            .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn test_create() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_create_user()
-            .withf(|_, req: &UserCreate| req.domain_id == "domain" && req.name == "name")
-            .returning(|_, req| {
-                Ok(UserResponseBuilder::default()
-                    .id("bar")
-                    .domain_id(req.domain_id.clone())
-                    .enabled(true)
-                    .name(req.name.clone())
-                    .build()
-                    .unwrap())
-            });
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state.clone());
-
-        let user = UserCreateRequest {
-            user: ApiUserCreate::default()
-                .domain_id("domain")
-                .name("name")
-                .build()
-                .unwrap(),
-        };
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .header(http::header::CONTENT_TYPE, "application/json")
-                    .uri("/")
-                    .extension(vsc)
-                    .body(Body::from(serde_json::to_string(&user).unwrap()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::CREATED);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let created_user: ApiUserResponse = serde_json::from_slice(&body).unwrap();
-        assert_eq!(created_user.user.name, user.user.name);
-    }
-
-    #[tokio::test]
-    #[tracing_test::traced_test]
-    async fn test_get() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_get_user()
-            .withf(|_, id: &'_ str| id == "foo")
-            .returning(|_, _| Ok(None));
-
-        identity_mock
-            .expect_get_user()
-            .withf(|_, id: &'_ str| id == "bar")
-            .returning(|_, _| {
-                Ok(Some(
-                    UserResponseBuilder::default()
-                        .id("bar")
-                        .domain_id("did")
-                        .enabled(true)
-                        .name("foo")
-                        .build()
-                        .unwrap(),
-                ))
-            });
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state.clone());
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .uri("/foo")
-                    .extension(vsc.clone())
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::NOT_FOUND);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .uri("/bar")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let res: ApiUserResponse = serde_json::from_slice(&body).unwrap();
-        assert_eq!(
-            ApiUser::default()
-                .id("bar")
-                .domain_id("did")
-                .enabled(true)
-                .name("foo")
-                .build()
-                .unwrap(),
-            res.user,
-        );
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_delete_user()
-            .withf(|_ctx: &ExecutionContext<'_>, id: &'_ str| id == "foo")
-            .returning(|_, _| Err(IdentityProviderError::UserNotFound("foo".into())));
-
-        identity_mock
-            .expect_delete_user()
-            .withf(|_ctx: &ExecutionContext<'_>, id: &'_ str| id == "bar")
-            .returning(|_, _| Ok(()));
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state.clone());
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri("/foo")
-                    .extension(vsc.clone())
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::NOT_FOUND);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri("/bar")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::NO_CONTENT);
-    }
-
-    #[tokio::test]
-    async fn test_groups() {
-        let vsc = test_fixture_scoped();
-        let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_list_groups_of_user()
-            .withf(|_, uid: &str| uid == "foo")
-            .returning(|_, _| {
-                Ok(vec![Group {
-                    id: "1".into(),
-                    name: "2".into(),
-                    domain_id: "did".into(),
-                    ..Default::default()
-                }])
-            });
-
-        let state = get_mocked_state(
-            Provider::mocked_builder().mock_identity(identity_mock),
-            true,
-            None,
-        )
-        .await;
-
-        let mut api = openapi_router()
-            .layer(TraceLayer::new_for_http())
-            .with_state(state);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .uri("/foo/groups")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let res: GroupList = serde_json::from_slice(&body).unwrap();
-        assert_eq!(
-            vec![
-                ApiGroupBuilder::default()
-                    .id("1")
-                    .name("2")
-                    .domain_id("did")
-                    .build()
-                    .unwrap()
-            ],
-            res.groups
-        );
-    }
+        .routes(routes!(list::list, create::create))
+        .routes(routes!(show::show, delete::remove))
+        .routes(routes!(update::update))
+        .routes(routes!(groups::groups))
 }
