+++ crates/keystone/src/api/v4/user/update.rs
@@ -11,60 +11,79 @@
 // limitations under the License.
 //
 // SPDX-License-Identifier: Apache-2.0
+//! User: update.
 
 use axum::{
+    Json,
     extract::{Path, State},
     http::StatusCode,
     response::IntoResponse,
 };
 use serde_json::json;
+use validator::Validate;
 
+use super::types::{User, UserResponse, UserUpdateRequest};
 use crate::api::auth::Auth;
 use crate::api::error::KeystoneApiError;
 use crate::keystone::ServiceState;
 use openstack_keystone_core::auth::ExecutionContext;
 
-/// Delete user
+/// Update existing user
 #[utoipa::path(
-    delete,
+    put,
     path = "/{user_id}",
-    description = "Delete user by ID",
+    description = "Update user by ID",
     params(),
     responses(
-        (status = 204, description = "Deleted"),
+        (status = OK, description = "Updated user", body = UserResponse),
         (status = 404, description = "User not found", example = json!(KeystoneApiError::NotFound(String::from("id = 1"))))
     ),
     tag="users"
 )]
-#[tracing::instrument(name = "api::user_delete", level = "debug", skip(state))]
-pub(super) async fn delete(
+#[tracing::instrument(name = "api::update_user", level = "debug", skip(state))]
+pub(super) async fn update(
     Auth(user_auth): Auth,
     Path(user_id): Path<String>,
     State(state): State<ServiceState>,
+    Json(req): Json<UserUpdateRequest>,
 ) -> Result<impl IntoResponse, KeystoneApiError> {
+    req.validate()?;
+
     let current = state
         .provider
         .get_identity_provider()
         .get_user(&ExecutionContext::from_auth(&state, &user_auth), &user_id)
         .await?;
 
+    let existing_user = current.as_ref().map(|c| json!({"user": c}));
+
     state
         .policy_enforcer
         .enforce(
-            "identity/user/delete",
+            "identity/user/update",
             &user_auth,
-            serde_json::Value::Null,
-            Some(json!({"user": current})),
+            json!({"user": req.user}),
+            existing_user,
         )
         .await?;
+
     match current {
         Some(_) => {
-            state
+            let user = state
                 .provider
                 .get_identity_provider()
-                .delete_user(&ExecutionContext::from_auth(&state, &user_auth), &user_id)
+                .update_user(
+                    &ExecutionContext::from_auth(&state, &user_auth),
+                    &user_id,
+                    req.into(),
+                )
                 .await?;
-            Ok((StatusCode::NO_CONTENT).into_response())
+            Ok((
+                StatusCode::OK,
+                Json(UserResponse {
+                    user: User::from(user),
+                }),
+            ))
         }
         _ => Err(KeystoneApiError::NotFound {
             resource: "user".to_string(),
@@ -77,25 +96,26 @@ pub(super) async fn delete(
 mod tests {
     use axum::{
         body::Body,
-        http::{Request, StatusCode},
+        http::{self, Request, StatusCode},
     };
-    use tower::ServiceExt; // for `call`, `oneshot`, and `ready`
+    use http_body_util::BodyExt;
+    use tower::ServiceExt;
     use tower_http::trace::TraceLayer;
+
+    use openstack_keystone_core_types::identity::*;
 
     use super::super::openapi_router;
     use crate::api::tests::{get_mocked_state, test_fixture_scoped};
+    use crate::api::v3::user::types::{
+        UserResponse as ApiUserResponse, UserUpdateBuilder as ApiUserUpdate, UserUpdateRequest,
+    };
     use crate::identity::MockIdentityProvider;
     use crate::provider::Provider;
-    use openstack_keystone_core::auth::ExecutionContext;
-    use openstack_keystone_core_types::identity::UserResponseBuilder;
 
     #[tokio::test]
-    async fn test_delete() {
+    async fn test_update() {
+        let vsc = test_fixture_scoped();
         let mut identity_mock = MockIdentityProvider::default();
-        identity_mock
-            .expect_get_user()
-            .withf(|_, id: &'_ str| id == "foo")
-            .returning(|_, _| Ok(None));
 
         identity_mock
             .expect_get_user()
@@ -104,20 +124,31 @@ mod tests {
                 Ok(Some(
                     UserResponseBuilder::default()
                         .id("bar")
-                        .domain_id("user_domain_id")
+                        .domain_id("did")
                         .enabled(true)
-                        .name("name")
+                        .name("old_name")
                         .build()
                         .unwrap(),
                 ))
             });
 
         identity_mock
-            .expect_delete_user()
-            .withf(|_ctx: &ExecutionContext<'_>, id: &'_ str| id == "bar")
-            .returning(|_, _| Ok(()));
+            .expect_update_user()
+            .withf(
+                |_, id: &'_ str, _: &openstack_keystone_core_types::identity::UserUpdate| {
+                    id == "bar"
+                },
+            )
+            .returning(|_, _, _| {
+                Ok(UserResponseBuilder::default()
+                    .id("bar")
+                    .domain_id("did")
+                    .enabled(true)
+                    .name("new_name")
+                    .build()
+                    .unwrap())
+            });
 
-        let vsc = test_fixture_scoped();
         let state = get_mocked_state(
             Provider::mocked_builder().mock_identity(identity_mock),
             true,
@@ -129,41 +160,79 @@ mod tests {
             .layer(TraceLayer::new_for_http())
             .with_state(state.clone());
 
+        let update_req = UserUpdateRequest {
+            user: ApiUserUpdate::default().name("new_name").build().unwrap(),
+        };
+
         let response = api
             .as_service()
             .oneshot(
                 Request::builder()
-                    .method("DELETE")
-                    .uri("/foo")
-                    .extension(vsc.clone())
-                    .body(Body::empty())
+                    .method("PUT")
+                    .header(http::header::CONTENT_TYPE, "application/json")
+                    .uri("/bar")
+                    .extension(vsc)
+                    .body(Body::from(serde_json::to_string(&update_req).unwrap()))
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+
+        let body = response.into_body().collect().await.unwrap().to_bytes();
+        let updated_user: ApiUserResponse = serde_json::from_slice(&body).unwrap();
+        assert_eq!(updated_user.user.name, "new_name");
+        assert_eq!(updated_user.user.id, "bar");
+    }
+
+    #[tokio::test]
+    async fn test_update_not_found() {
+        let vsc = test_fixture_scoped();
+        let mut identity_mock = MockIdentityProvider::default();
+
+        identity_mock
+            .expect_get_user()
+            .withf(|_, id: &'_ str| id == "missing")
+            .returning(|_, _| Ok(None));
+
+        let state = get_mocked_state(
+            Provider::mocked_builder().mock_identity(identity_mock),
+            true,
+            None,
+        )
+        .await;
+
+        let mut api = openapi_router()
+            .layer(TraceLayer::new_for_http())
+            .with_state(state.clone());
+
+        let update_req = UserUpdateRequest {
+            user: ApiUserUpdate::default().name("new_name").build().unwrap(),
+        };
+
+        let response = api
+            .as_service()
+            .oneshot(
+                Request::builder()
+                    .method("PUT")
+                    .header(http::header::CONTENT_TYPE, "application/json")
+                    .uri("/missing")
+                    .extension(vsc)
+                    .body(Body::from(serde_json::to_string(&update_req).unwrap()))
                     .unwrap(),
             )
             .await
             .unwrap();
 
         assert_eq!(response.status(), StatusCode::NOT_FOUND);
-
-        let response = api
-            .as_service()
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri("/bar")
-                    .extension(vsc)
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::NO_CONTENT);
     }
 
     #[tokio::test]
-    async fn test_delete_policy_denied() {
+    async fn test_update_policy_denied() {
         let vsc = test_fixture_scoped();
         let mut identity_mock = MockIdentityProvider::default();
+
         identity_mock
             .expect_get_user()
             .withf(|_, id: &'_ str| id == "bar")
@@ -171,9 +240,9 @@ mod tests {
                 Ok(Some(
                     UserResponseBuilder::default()
                         .id("bar")
-                        .domain_id("user_domain_id")
+                        .domain_id("did")
                         .enabled(true)
-                        .name("name")
+                        .name("old_name")
                         .build()
                         .unwrap(),
                 ))
@@ -190,14 +259,19 @@ mod tests {
             .layer(TraceLayer::new_for_http())
             .with_state(state.clone());
 
+        let update_req = UserUpdateRequest {
+            user: ApiUserUpdate::default().name("new_name").build().unwrap(),
+        };
+
         let response = api
             .as_service()
             .oneshot(
                 Request::builder()
-                    .method("DELETE")
+                    .method("PUT")
+                    .header(http::header::CONTENT_TYPE, "application/json")
                     .uri("/bar")
                     .extension(vsc)
-                    .body(Body::empty())
+                    .body(Body::from(serde_json::to_string(&update_req).unwrap()))
                     .unwrap(),
             )
             .await
@@ -207,20 +281,25 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn test_delete_unauth() {
+    async fn test_update_unauth() {
         let state = get_mocked_state(Provider::mocked_builder(), false, None).await;
 
         let mut api = openapi_router()
             .layer(TraceLayer::new_for_http())
-            .with_state(state);
+            .with_state(state.clone());
+
+        let update_req = UserUpdateRequest {
+            user: ApiUserUpdate::default().name("new_name").build().unwrap(),
+        };
 
         let response = api
             .as_service()
             .oneshot(
                 Request::builder()
-                    .method("DELETE")
+                    .method("PUT")
+                    .header(http::header::CONTENT_TYPE, "application/json")
                     .uri("/bar")
-                    .body(Body::empty())
+                    .body(Body::from(serde_json::to_string(&update_req).unwrap()))
                     .unwrap(),
             )
             .await
