+++ policy/identity/group/show_test.rego
@@ -4,12 +4,12 @@ import data.identity.group.show
 
 test_allowed if {
 	show.allow with input as {"credentials": {"roles": ["admin"]}}
-	show.allow with input as {"credentials": {"roles": ["manager", "reader"], "domain_id": "foo"}, "target": {"group": {"domain_id": "foo"}}}
+	show.allow with input as {"credentials": {"roles": ["manager", "reader"], "domain_id": "foo"}, "existing": {"group": {"domain_id": "foo"}}}
 }
 
 test_forbidden if {
 	not show.allow with input as {"credentials": {"roles": []}}
-	not show.allow with input as {"credentials": {"roles": ["manager"], "domain_id": "foo"}, "target": {"group": {"domain_id": "foo1"}}}
-	not show.allow with input as {"credentials": {"roles": ["manager"]}, "target": {"group": {"domain_id": "foo"}}}
-	not show.allow with input as {"credentials": {"roles": ["reader"], "domain_id": "foo"}, "target": {"group": {"domain_id": "foo2"}}}
+	not show.allow with input as {"credentials": {"roles": ["manager"], "domain_id": "foo"}, "existing": {"group": {"domain_id": "foo1"}}}
+	not show.allow with input as {"credentials": {"roles": ["manager"]}, "existing": {"group": {"domain_id": "foo"}}}
+	not show.allow with input as {"credentials": {"roles": ["reader"], "domain_id": "foo"}, "existing": {"group": {"domain_id": "foo2"}}}
 }
