+++ policy/mapping/ruleset/create.rego
@@ -25,46 +25,46 @@ import data.identity.mapping as mapping_common
 default allow := false
 
 allow if {
-    not mapping_common.global_ruleset
-    mapping_common.own_ruleset
-    not mapping_common.is_system_ruleset
-    not mapping_common.claims_mode_resolution
-    "manager" in input.credentials.roles
+	not mapping_common.global_ruleset
+	mapping_common.own_ruleset
+	not mapping_common.is_system_ruleset
+	not mapping_common.claims_mode_resolution
+	"manager" in input.credentials.roles
 }
 
 allow if {
-    "admin" in input.credentials.roles
+	"admin" in input.credentials.roles
 }
 
 violation contains {"field": "domain_id", "msg": msg} if {
-    mapping_common.global_ruleset
-    not "admin" in input.credentials.roles
-    msg := "creating a global mapping ruleset requires `admin` role."
+	mapping_common.global_ruleset
+	not "admin" in input.credentials.roles
+	msg := "creating a global mapping ruleset requires `admin` role."
 }
 
 violation contains {"field": "domain_id", "msg": msg} if {
-    mapping_common.foreign_ruleset
-    not "admin" in input.credentials.roles
-    msg := "creating a mapping ruleset for another domain requires `admin` role."
+	mapping_common.foreign_ruleset
+	not "admin" in input.credentials.roles
+	msg := "creating a mapping ruleset for another domain requires `admin` role."
 }
 
 violation contains {"field": "is_system", "msg": msg} if {
-    mapping_common.is_system_ruleset
-    not "admin" in input.credentials.roles
-    msg := "creating a system mapping ruleset requires `admin` role."
+	mapping_common.is_system_ruleset
+	not "admin" in input.credentials.roles
+	msg := "creating a system mapping ruleset requires `admin` role."
 }
 
 violation contains {"field": "domain_resolution_mode", "msg": msg} if {
-    mapping_common.claims_mode_resolution
-    not "admin" in input.credentials.roles
-    msg := "claims-based domain resolution mode requires `admin` role."
+	mapping_common.claims_mode_resolution
+	not "admin" in input.credentials.roles
+	msg := "claims-based domain resolution mode requires `admin` role."
 }
 
 violation contains {"field": "role", "msg": msg} if {
-    mapping_common.own_ruleset
-    not mapping_common.is_system_ruleset
-    not mapping_common.claims_mode_resolution
-    not "admin" in input.credentials.roles
-    not "manager" in input.credentials.roles
-    msg := "creating a mapping ruleset requires `manager` role."
+	mapping_common.own_ruleset
+	not mapping_common.is_system_ruleset
+	not mapping_common.claims_mode_resolution
+	not "admin" in input.credentials.roles
+	not "manager" in input.credentials.roles
+	msg := "creating a mapping ruleset requires `manager` role."
 }
